Set up benchmarks

+++ inc/EigenPath.hpp
@@ -90,12 +90,12 @@ class EigenPath {
     explicit EigenPath(const RobotPtr& robot, const ompl::base::PathPtr& path);
 
     /*\brief Get configuration along path at position s in [0,1] */
-    StateXd GetConfigAt(const float position) const;
+    StateXd GetConfigAt(const double position) const;
 
-    float GetLength() const;
+    double GetLength() const;
 
-    float GetStartTime() const;
-    float GetEndTime() const;
+    double GetStartTime() const;
+    double GetEndTime() const;
 
     void InitLengthFromConfigs(const std::vector<StateXd>& configs);
 
@@ -105,10 +105,9 @@ class EigenPath {
 
   private:
     std::vector<StateXd> configs_;
-    std::vector<float> time_at_configs_;
-    //std::vector<float> lengths_;
-    float total_length_{0.0f};
+    std::vector<double> time_at_configs_;
+    double total_length_{0.0f};
 
-    float start_time_{0.0};
-    float end_time_{1.0};
+    double start_time_{0.0};
+    double end_time_{1.0};
 };
